fix type error caused with VSCode only
